+++ packages/svelte/src/lib/plugins/NodeResizer/index.ts
@@ -1,0 +1,2 @@
+export { default as NodeResizer } from './NodeResizer.svelte';
+export { default as NodeResizeControl } from './ResizeControl.svelte';