+++ packages/system/src/index.ts
@@ -5,3 +5,4 @@ export * from './xydrag';
 export * from './xyhandle';
 export * from './xyminimap';
 export * from './xypanzoom';
+export * from './xyresizer';