+++ packages/system/src/types/utils.ts
@@ -1,5 +1,3 @@
-import type { Optional } from '../utils/types';
-
 export enum Position {
   Left = 'left',
   Top = 'top',