+++ packages/react/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Minor changes
 
-- fix applyChanges: handle multi changes for one node
+- fix applyChanges: handle multi changes for one node, deletions and expandParent
 - use `XYResizer` from @xyflow/system
+- add unit tests for `applyNodeChanges` and `applyEdgeChanges`
 
 ## 12.0.0-next.4